@@ -1,152 +1,149 @@
 {
-    "name": "pulp-fiction",
-    "version": "0.0.0",
-    "license": "MIT",
-    "scripts": {
-        "nx": "nx",
-        "start": "nx serve",
-        "build": "nx build",
-        "build:ci:server": "nx build server --skip-nx-cache --prod",
-        "build:ci:client": "nx build client --skip-nx-cache --prod",
-        "test": "nx test",
-        "lint": "nx workspace-lint && nx lint",
-        "e2e": "nx e2e",
-        "affected:apps": "nx affected:apps",
-        "affected:libs": "nx affected:libs",
-        "affected:build": "nx affected:build",
-        "affected:e2e": "nx affected:e2e",
-        "affected:test": "nx affected:test",
-        "affected:lint": "nx affected:lint",
-        "affected:dep-graph": "nx affected:dep-graph",
-        "affected": "nx affected",
-        "format": "nx format:write",
-        "format:write": "nx format:write",
-        "format:check": "nx format:check",
-        "update": "nx migrate latest",
-        "workspace-schematic": "nx workspace-schematic",
-        "dep-graph": "nx dep-graph",
-        "help": "nx help",
-        "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular-devkit/build-angular": "0.1100.5",
-        "@angular/animations": "11.0.5",
-        "@angular/cdk": "^10.2.0",
-        "@angular/common": "11.0.5",
-        "@angular/compiler": "11.0.5",
-        "@angular/compiler-cli": "11.0.5",
-        "@angular/core": "11.0.5",
-        "@angular/flex-layout": "^10.0.0-beta.32",
-        "@angular/forms": "11.0.5",
-        "@angular/language-service": "11.0.5",
-        "@angular/material": "^10.2.0",
-        "@angular/platform-browser": "11.0.5",
-        "@angular/platform-browser-dynamic": "11.0.5",
-        "@angular/router": "11.0.5",
-        "@auth0/angular-jwt": "^5.0.1",
-        "@ckeditor/ckeditor5-angular": "^1.2.3",
-        "@ckeditor/ckeditor5-indent": "^22.0.0",
-        "@ckeditor/ckeditor5-ui": "^22.0.0",
-        "@nestjs/common": "^7.6.5",
-        "@nestjs/core": "^7.6.5",
-        "@nestjs/graphql": "^7.6.0",
-        "@nestjs/jwt": "^7.1.0",
-        "@nestjs/mongoose": "^7.0.2",
-        "@nestjs/platform-express": "^7.0.0",
-        "@nestjs/platform-socket.io": "^7.4.4",
-        "@nestjs/serve-static": "^2.1.4",
-        "@nestjs/swagger": "^4.7.12",
-        "@nestjs/websockets": "^7.4.4",
-        "@nestjsplus/cookies": "^1.1.1",
-        "@ngneat/until-destroy": "^8.0.3",
-        "@ngu/carousel": "^3.0.2",
-        "@ngx-loading-bar/core": "^5.1.1",
-        "@ngx-loading-bar/http-client": "^5.1.1",
-        "@ngxs-labs/dispatch-decorator": "^2.1.0",
-        "@ngxs-labs/select-snapshot": "^2.0.1",
-        "@ngxs/devtools-plugin": "^3.7.1",
-        "@ngxs/logger-plugin": "^3.7.1",
-        "@ngxs/router-plugin": "^3.7.1",
-        "@ngxs/storage-plugin": "^3.7.1",
-        "@ngxs/store": "^3.7.1",
-        "@nrwl/angular": "11.2.10",
-        "angular-feather": "^6.0.10",
-        "argon2": "^0.27.0",
-        "aws-sdk": "^2.738.0",
-        "body-parser": "^1.19.0",
-        "cookie-parser": "^1.4.5",
-        "csurf": "^1.11.0",
-        "graphql": "^15.3.0",
-        "hammerjs": "^2.0.8",
-        "helmet": "^4.4.1",
-        "jsonwebtoken": "^8.5.1",
-        "lodash": "^4.17.20",
-        "markdown-it": "^11.0.0",
-        "mongoose": "^5.10.0",
-        "mongoose-autopopulate": "^0.12.2",
-        "mongoose-paginate-v2": "^1.3.9",
-        "nanoid": "^3.1.20",
-        "ng-event-source": "^1.0.14",
-        "ng2-file-upload": "^1.4.0",
-        "ng2-fittext": "^1.2.12",
-        "ngx-cookie": "^5.0.0",
-        "ngx-image-cropper": "^3.2.1",
-        "ngx-markdown": "^11.0.1",
-        "ngx-pagination": "^5.0.0",
-        "numeral": "^2.0.6",
-        "prerender-node": "^3.2.5",
-        "reflect-metadata": "^0.1.13",
-        "rxjs": "~6.5.5",
-        "sanitize-html": "^2.3.2",
-        "swagger-ui-express": "^4.1.6",
-        "tslib": "^2.0.0",
-        "validator": "^13.1.1",
-        "voca": "^1.4.0",
-        "zone.js": "^0.10.2"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~0.1100.1",
-        "@angular/compiler-cli": "^11.0.0",
-        "@angular/language-service": "^11.0.0",
-        "@nestjs/schematics": "^7.0.0",
-        "@nestjs/testing": "^7.0.0",
-        "@nrwl/cypress": "11.2.10",
-        "@nrwl/eslint-plugin-nx": "11.2.10",
-        "@nrwl/jest": "11.2.10",
-        "@nrwl/nest": "11.2.10",
-        "@nrwl/node": "11.2.10",
-        "@nrwl/workspace": "11.2.10",
-        "@types/aws-sdk": "^2.7.0",
-        "@types/cookie-parser": "^1.4.2",
-        "@types/csurf": "^1.11.0",
-        "@types/dotenv": "^8.2.0",
-        "@types/graphql": "^14.5.0",
-        "@types/helmet": "^4.0.0",
-        "@types/jest": "26.0.8",
-        "@types/lodash": "^4.14.159",
-        "@types/markdown-it": "^10.0.2",
-        "@types/mongoose": "^5.7.36",
-        "@types/mongoose-paginate-v2": "^1.3.8",
-        "@types/node": "12.12.38",
-        "@types/sanitize-html": "^1.27.1",
-        "@types/socket.io": "^2.1.11",
-        "@types/uuid": "^8.3.0",
-        "@types/voca": "^1.4.0",
-        "@typescript-eslint/eslint-plugin": "4.3.0",
-        "@typescript-eslint/parser": "4.3.0",
-        "@wasm-tool/wasm-pack-plugin": "^1.3.1",
-        "codelyzer": "6.0.1",
-        "cypress": "^4.1.0",
-        "dotenv": "^8.2.0",
-        "eslint": "7.10.0",
-        "eslint-config-prettier": "6.0.0",
-        "jest": "26.2.2",
-        "jest-preset-angular": "8.3.2",
-        "prettier": "2.2.1",
-        "ts-jest": "26.4.0",
-        "ts-node": "9.1.1",
-        "tslint": "6.1.3",
-        "typescript": "4.0.5"
-    }
+  "name": "pulp-fiction",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "nx": "nx",
+    "start": "nx serve",
+    "build": "nx build",
+    "build:ci:server": "nx build server --skip-nx-cache --prod",
+    "build:ci:client": "nx build client --skip-nx-cache --prod",
+    "test": "nx test",
+    "lint": "nx workspace-lint && nx lint",
+    "e2e": "nx e2e",
+    "affected:apps": "nx affected:apps",
+    "affected:libs": "nx affected:libs",
+    "affected:build": "nx affected:build",
+    "affected:e2e": "nx affected:e2e",
+    "affected:test": "nx affected:test",
+    "affected:lint": "nx affected:lint",
+    "affected:dep-graph": "nx affected:dep-graph",
+    "affected": "nx affected",
+    "format": "nx format:write",
+    "format:write": "nx format:write",
+    "format:check": "nx format:check",
+    "update": "nx migrate latest",
+    "workspace-schematic": "nx workspace-schematic",
+    "dep-graph": "nx dep-graph",
+    "help": "nx help",
+    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular-devkit/build-angular": "~0.1100.1",
+    "@angular/animations": "^11.2.0",
+    "@angular/cdk": "^11.2.0",
+    "@angular/common": "^11.2.0",
+    "@angular/compiler": "^11.2.0",
+    "@angular/core": "^11.2.0",
+    "@angular/flex-layout": "^10.0.0-beta.32",
+    "@angular/forms": "^11.2.0",
+    "@angular/material": "^11.2.0",
+    "@angular/platform-browser": "^11.2.0",
+    "@angular/platform-browser-dynamic": "^11.2.0",
+    "@angular/router": "^11.2.0",
+    "@auth0/angular-jwt": "^5.0.1",
+    "@ckeditor/ckeditor5-angular": "^2.0.1",
+    "@ckeditor/ckeditor5-indent": "^25.0.0",
+    "@ckeditor/ckeditor5-ui": "^25.0.0",
+    "@nestjs/common": "^7.6.5",
+    "@nestjs/core": "^7.6.5",
+    "@nestjs/graphql": "^7.6.0",
+    "@nestjs/jwt": "^7.1.0",
+    "@nestjs/mongoose": "^7.0.2",
+    "@nestjs/platform-express": "^7.0.0",
+    "@nestjs/platform-socket.io": "^7.4.4",
+    "@nestjs/serve-static": "^2.1.4",
+    "@nestjs/swagger": "^4.7.12",
+    "@nestjs/websockets": "^7.4.4",
+    "@nestjsplus/cookies": "^1.1.1",
+    "@ngneat/until-destroy": "^8.0.3",
+    "@ngu/carousel": "^3.0.2",
+    "@ngx-loading-bar/core": "^5.1.1",
+    "@ngx-loading-bar/http-client": "^5.1.1",
+    "@ngxs-labs/dispatch-decorator": "^2.1.0",
+    "@ngxs-labs/select-snapshot": "^2.0.1",
+    "@ngxs/devtools-plugin": "^3.7.1",
+    "@ngxs/logger-plugin": "^3.7.1",
+    "@ngxs/router-plugin": "^3.7.1",
+    "@ngxs/storage-plugin": "^3.7.1",
+    "@ngxs/store": "^3.7.1",
+    "@nrwl/angular": "11.2.12",
+    "angular-feather": "^6.0.10",
+    "argon2": "^0.27.0",
+    "aws-sdk": "^2.738.0",
+    "body-parser": "^1.19.0",
+    "cookie-parser": "^1.4.5",
+    "csurf": "^1.11.0",
+    "graphql": "^15.3.0",
+    "hammerjs": "^2.0.8",
+    "helmet": "^4.4.1",
+    "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.20",
+    "markdown-it": "^11.0.0",
+    "mongoose": "^5.10.0",
+    "mongoose-autopopulate": "^0.12.2",
+    "mongoose-paginate-v2": "^1.3.9",
+    "nanoid": "^3.1.20",
+    "ng-event-source": "^1.0.14",
+    "ng2-file-upload": "^1.4.0",
+    "ng2-fittext": "^1.2.12",
+    "ngx-cookie": "^5.0.0",
+    "ngx-image-cropper": "^3.2.1",
+    "ngx-markdown": "^11.0.1",
+    "ngx-pagination": "^5.0.0",
+    "numeral": "^2.0.6",
+    "prerender-node": "^3.2.5",
+    "reflect-metadata": "^0.1.13",
+    "rxjs": "~6.5.5",
+    "sanitize-html": "^2.3.2",
+    "swagger-ui-express": "^4.1.6",
+    "tslib": "^2.0.0",
+    "validator": "^13.1.1",
+    "voca": "^1.4.0",
+    "zone.js": "^0.11.3"
+  },
+  "devDependencies": {
+    "@angular/compiler-cli": "^11.0.0",
+    "@angular/language-service": "^11.0.0",
+    "@nestjs/schematics": "^7.0.0",
+    "@nestjs/testing": "^7.0.0",
+    "@nrwl/cypress": "11.2.12",
+    "@nrwl/eslint-plugin-nx": "11.2.12",
+    "@nrwl/jest": "11.2.12",
+    "@nrwl/nest": "11.2.12",
+    "@nrwl/node": "11.2.12",
+    "@nrwl/workspace": "11.2.12",
+    "@types/aws-sdk": "^2.7.0",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/csurf": "^1.11.0",
+    "@types/dotenv": "^8.2.0",
+    "@types/graphql": "^14.5.0",
+    "@types/helmet": "^4.0.0",
+    "@types/jest": "26.0.8",
+    "@types/lodash": "^4.14.159",
+    "@types/markdown-it": "^10.0.2",
+    "@types/mongoose": "^5.7.36",
+    "@types/mongoose-paginate-v2": "^1.3.8",
+    "@types/node": "12.12.38",
+    "@types/sanitize-html": "^1.27.1",
+    "@types/socket.io": "^2.1.11",
+    "@types/uuid": "^8.3.0",
+    "@types/voca": "^1.4.0",
+    "@typescript-eslint/eslint-plugin": "4.3.0",
+    "@typescript-eslint/parser": "4.3.0",
+    "@wasm-tool/wasm-pack-plugin": "^1.3.1",
+    "codelyzer": "6.0.1",
+    "cypress": "^4.1.0",
+    "dotenv": "^8.2.0",
+    "eslint": "7.10.0",
+    "eslint-config-prettier": "6.0.0",
+    "jest": "26.2.2",
+    "jest-preset-angular": "8.3.2",
+    "prettier": "2.2.1",
+    "ts-jest": "26.4.0",
+    "ts-node": "9.1.1",
+    "tslint": "6.1.3",
+    "typescript": "^4.1.5"
+  }
 }
