@@ -10,18 +10,32 @@ import { AlertsModule } from '@dragonfish/alerts';
 import { PipesModule } from '@dragonfish/pipes';
 import { EditorModule } from '@dragonfish/editor';
 import { ComponentsModule } from '@dragonfish/components';
-
 import { MyStuffRoutingModule } from './my-stuff-routing.module';
+
+/* State */
 import { MyStuffState } from './shared/my-stuff.state';
 import { SectionsState } from './shared/sections';
-import { MyStuffComponents } from './components';
-import { MyStuffViews } from './views';
+
+/* Components */
+import { ContentItemComponent, SectionItemComponent, UploadCoverartComponent } from './components';
+import { BlogFormComponent, NewsFormComponent, PoetryFormComponent, ProseFormComponent, ViewPoetryComponent, ViewProseComponent } from './views';
 import { MyStuffFacade } from './facades';
 import { MyStuffService } from './shared/services';
 import { MyStuffComponent } from './my-stuff.component';
 
 @NgModule({
-    declarations: [MyStuffComponent, ...MyStuffComponents, ...MyStuffViews],
+    declarations: [
+        MyStuffComponent, 
+        BlogFormComponent,
+        NewsFormComponent,
+        PoetryFormComponent,
+        ProseFormComponent,
+        ViewProseComponent,
+        ViewPoetryComponent,
+        ContentItemComponent,
+        SectionItemComponent, 
+        UploadCoverartComponent,
+    ],
     imports: [
         CommonModule,
         MyStuffRoutingModule,
