@@ -1,4 +1,2 @@
-import { MyStuffResolver } from './my-stuff.resolver';
-import { ViewSectionsResolver } from './view-sections.resolver';
-
-export const MyStuffResolvers = [MyStuffResolver, ViewSectionsResolver];
+export { MyStuffResolver } from './my-stuff.resolver';
+export { ViewSectionsResolver } from './view-sections.resolver';
