@@ -1,15 +1,6 @@
-import { BlogFormComponent } from './blog-form/blog-form.component';
-import { NewsFormComponent } from './news-form/news-form.component';
-import { PoetryFormComponent } from './poetry-form/poetry-form.component';
-import { ProseFormComponent } from './prose-form/prose-form.component';
-import { ViewPoetryComponent } from './view-poetry/view-poetry.component';
-import { ViewProseComponent } from './view-prose/view-prose.component';
-
-export const MyStuffViews = [
-    BlogFormComponent,
-    NewsFormComponent,
-    PoetryFormComponent,
-    ProseFormComponent,
-    ViewPoetryComponent,
-    ViewProseComponent,
-];
+export { BlogFormComponent } from './blog-form/blog-form.component';
+export { NewsFormComponent } from './news-form/news-form.component';
+export { PoetryFormComponent } from './poetry-form/poetry-form.component';
+export { ProseFormComponent } from './prose-form/prose-form.component';
+export { ViewPoetryComponent } from './view-poetry/view-poetry.component';
+export { ViewProseComponent } from './view-prose/view-prose.component';
