@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 /* Resolvers */
-import { DashboardResolvers } from './resolvers/dashboard';
+import { BrowseResolvers } from './resolvers/browse';
 import { DocsResolvers } from './resolvers/docs';
 import { HomeResolvers } from './resolvers/home';
 import { MigrationResolvers } from './resolvers/migration';
@@ -11,33 +11,40 @@ import { Resolvers } from './resolvers';
 
 /* Page Routes */
 import { HomeRoutes } from './pages/home';
-import { BrowseRoutes } from './pages/browse';
-import { SocialRoutes } from './pages/social';
-import { ContentViewRoutes } from './pages/content-views';
 import { AccountRoutes } from './pages/account';
-import { PortfolioRoutes } from './pages/portfolio';
-import { SearchRoutes } from './pages/search';
+import { BrowseRoutes } from './pages/browse';
+import { ContentViewRoutes } from './pages/content-views';
 import { DocsRoutes } from './pages/docs';
 import { MigrationRoutes } from './pages/migration';
-import { DashboardRoutes } from './pages/dash';
+import { SocialRoutes } from './pages/social';
+import { SearchRoutes } from './pages/search';
+import { PortfolioRoutes } from './pages/portfolio';
+
+/* Util */
+import { Roles } from '@dragonfish/models/users';
 import { AuthGuard } from './shared/auth/services';
 
 const routes: Routes = [
     { path: '', redirectTo: '/home', pathMatch: 'full' },
     ...HomeRoutes,
-    ...BrowseRoutes,
-    ...SocialRoutes,
-    ...ContentViewRoutes,
     ...AccountRoutes,
-    ...PortfolioRoutes,
-    ...SearchRoutes,
+    ...BrowseRoutes,
+    ...ContentViewRoutes,
     ...DocsRoutes,
     ...MigrationRoutes,
-    ...DashboardRoutes,
+    ...SocialRoutes,
+    ...SearchRoutes,
+    ...PortfolioRoutes,
     {
         path: 'my-stuff',
         canLoad: [AuthGuard],
         loadChildren: () => import('@dragonfish/my-stuff').then((m) => m.MyStuffModule),
+    },
+    {
+        path: 'dashboard',
+        canLoad: [AuthGuard],
+        data: { roles: [Roles.WorkApprover, Roles.Moderator, Roles.Admin] },
+        loadChildren: () => import('@dragonfish/dashboard').then((m) => m.DashboardModule),
     },
 ];
 
@@ -50,13 +57,6 @@ const routes: Routes = [
         }),
     ],
     exports: [RouterModule],
-    providers: [
-        ...DashboardResolvers,
-        ...DocsResolvers,
-        ...HomeResolvers,
-        ...MigrationResolvers,
-        ...PortfolioResolvers,
-        ...Resolvers,
-    ],
+    providers: [...DocsResolvers, ...HomeResolvers, ...MigrationResolvers, ...PortfolioResolvers, ...Resolvers, ...BrowseResolvers],
 })
 export class AppRoutingModule {}
