@@ -28,7 +28,6 @@ import { NgxsRouterPluginModule } from '@ngxs/router-plugin';
 import { AccountPages } from './pages/account';
 import { BrowsePages } from './pages/browse';
 import { ContentViewPages } from './pages/content-views';
-import { DashboardPages } from './pages/dash';
 import { DocsPages } from './pages/docs';
 import { HomePages } from './pages/home';
 import { MigrationPages } from './pages/migration';
@@ -41,7 +40,6 @@ import { AuthState } from './shared/auth';
 import { AuthInterceptor } from './shared/auth/services';
 import { GlobalState } from './shared/global';
 import { UserState } from './shared/user';
-import { ApprovalQueueState } from './shared/dash/approval-queue';
 import { ContentState } from './shared/content';
 
 /* Components */
@@ -80,7 +78,6 @@ import { environment } from '../environments/environment';
         ...AccountPages,
         ...BrowsePages,
         ...ContentViewPages,
-        ...DashboardPages,
         ...DocsPages,
         ...HomePages,
         ...MigrationPages,
@@ -125,7 +122,7 @@ import { environment } from '../environments/environment';
         PipesModule,
         EditorModule,
         ComponentsModule,
-        NgxsModule.forRoot([AuthState, GlobalState, UserState, ApprovalQueueState, ContentState], {
+        NgxsModule.forRoot([AuthState, GlobalState, UserState, ContentState], {
             developmentMode: !environment.production,
             selectorOptions: { suppressErrors: false, injectContainerState: false },
         }),
