@@ -76,7 +76,7 @@ export class UploadAvatarComponent implements OnInit {
         const mime = arr[0].match(/:(.*?);/)[1];
         const bstr = atob(arr[1]);
         let n = bstr.length;
-        let u8arr = new Uint8Array(n);
+        const u8arr = new Uint8Array(n);
 
         while (n--) {
             u8arr[n] = bstr.charCodeAt(n);
