@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 
 import { BrowseComponent } from './browse.component';
-import { Resolvers } from '../../resolvers';
+import { BrowseResolver } from '../../resolvers/browse';
 
 export const BrowsePages = [BrowseComponent];
 
@@ -9,11 +9,7 @@ export const BrowseRoutes: Routes = [
     {
         path: 'browse',
         component: BrowseComponent,
-        resolve: {
-            feedData: Resolvers.find((x) => {
-                return x.name === 'BrowseResolver';
-            }),
-        },
+        resolve: { feedData: BrowseResolver },
         runGuardsAndResolvers: 'always',
     },
 ];
