@@ -4,7 +4,7 @@ import { ProsePageComponent } from './prose-page/prose-page.component';
 import { PoetryPageComponent } from './poetry-page/poetry-page.component';
 import { PostPageComponent } from './post-page/post-page.component';
 import { SectionViewComponent } from './section-view/section-view.component';
-import { Resolvers } from '../../resolvers';
+import { ContentViewResolver } from '../../resolvers';
 
 export const ContentViewPages = [ProsePageComponent, PoetryPageComponent, PostPageComponent, SectionViewComponent];
 
@@ -12,9 +12,7 @@ export const ContentViewRoutes: Routes = [
     {
         path: 'post/:contentId/:postTitle',
         resolve: {
-            contentData: Resolvers.find((x) => {
-                return x.name === 'ContentViewResolver';
-            }),
+            contentData: ContentViewResolver,
         },
         runGuardsAndResolvers: 'paramsChange',
         component: PostPageComponent,
@@ -23,9 +21,7 @@ export const ContentViewRoutes: Routes = [
         path: 'prose/:contentId/:proseTitle',
         component: ProsePageComponent,
         resolve: {
-            contentData: Resolvers.find((x) => {
-                return x.name === 'ContentViewResolver';
-            }),
+            contentData: ContentViewResolver,
         },
         runGuardsAndResolvers: 'paramsChange',
         children: [{ path: ':sectionNum/:sectionTitle', component: SectionViewComponent }],
@@ -34,9 +30,7 @@ export const ContentViewRoutes: Routes = [
         path: 'poetry/:contentId/:poetryTitle',
         component: PoetryPageComponent,
         resolve: {
-            contentData: Resolvers.find((x) => {
-                return x.name === 'ContentViewResolver';
-            }),
+            contentData: ContentViewResolver,
         },
         runGuardsAndResolvers: 'paramsChange',
         children: [{ path: ':sectionNum/:sectionTitle', component: SectionViewComponent }],
