@@ -1,3 +1,1 @@
-import { ConversationsComponent as PortConversationsComponent } from './conversations.component';
-
-export const PortfolioConversationsPages = [PortConversationsComponent];
+export { ConversationsComponent as PortConversationsComponent } from './conversations.component';
