@@ -1,132 +1,102 @@
 import { Routes } from '@angular/router';
 
 import { PortfolioComponent } from './portfolio.component';
-import { PortfolioHomePages } from './port-home';
-import { PortfolioWorksPages } from './works';
-import { PortfolioBlogPages } from './blogs';
-import { PortfolioCollectionsPages } from './collections';
-import { PortfolioHistoryPages } from './history';
-import { PortfolioConversationsPages } from './conversations';
-import { PortfolioSettingsPages } from './settings';
+import { PortHomeComponent } from './port-home';
+import { WorksComponent } from './works';
+import { BlogPageComponent, BlogsComponent } from './blogs';
+import { CollectionPageComponent, CollectionsComponent } from './collections';
+import { PortHistComponent } from './history';
+import { PortConversationsComponent } from './conversations';
+import { SettingsComponent } from './settings';
 
 import { AuthGuard } from '../../shared/auth/services';
-import { Resolvers } from '../../resolvers';
-import { PortfolioResolvers } from '../../resolvers/portfolio';
+import { ContentViewResolver } from '../../resolvers';
+import { PortfolioResolver, BlogsResolver, WorksResolver, CollectionsResolver, CollectionPageResolver, HistoryResolver } from '../../resolvers/portfolio';
 
 export const PortfolioPages = [
     PortfolioComponent,
-    ...PortfolioHomePages,
-    ...PortfolioWorksPages,
-    ...PortfolioBlogPages,
-    ...PortfolioCollectionsPages,
-    ...PortfolioHistoryPages,
-    ...PortfolioConversationsPages,
-    ...PortfolioSettingsPages,
+    PortHomeComponent,
+    WorksComponent,
+    BlogsComponent,
+    BlogPageComponent,
+    CollectionsComponent,
+    CollectionPageComponent,
+    PortHistComponent,
+    PortConversationsComponent,
+    SettingsComponent,
 ];
 
 export const PortfolioRoutes: Routes = [
     {
         path: 'portfolio/:id/:username',
         resolve: {
-            portData: PortfolioResolvers.find((x) => {
-                return x.name === 'PortfolioResolver';
-            }),
+            portData: PortfolioResolver,
         },
         runGuardsAndResolvers: 'always',
         component: PortfolioComponent,
         children: [
             {
                 path: 'blogs',
-                component: PortfolioBlogPages.find((x) => {
-                    return x.name === 'BlogsComponent';
-                }),
+                component: BlogsComponent,
                 resolve: {
-                    feedData: PortfolioResolvers.find((x) => {
-                        return x.name === 'BlogsResolver';
-                    }),
+                    feedData: BlogsResolver,
                 },
                 runGuardsAndResolvers: 'always',
             },
             {
                 path: 'blog/:contentId',
                 resolve: {
-                    contentData: Resolvers.find((x) => {
-                        return x.name === 'ContentViewResolver';
-                    }),
+                    contentData: ContentViewResolver,
                 },
                 runGuardsAndResolvers: 'paramsChange',
-                component: PortfolioBlogPages.find((x) => {
-                    return x.name === 'BlogPageComponent';
-                }),
+                component: BlogPageComponent,
             },
             {
                 path: 'works',
-                component: PortfolioWorksPages.find((x) => {
-                    return x.name === 'WorksComponent';
-                }),
+                component: WorksComponent,
                 resolve: {
-                    feedData: PortfolioResolvers.find((x) => {
-                        return x.name === 'WorksResolver';
-                    }),
+                    feedData: WorksResolver,
                 },
                 runGuardsAndResolvers: 'always',
             },
             {
                 path: 'collections',
-                component: PortfolioCollectionsPages.find((x) => {
-                    return x.name === 'CollectionsComponent';
-                }),
+                component: CollectionsComponent,
                 resolve: {
-                    feedData: PortfolioResolvers.find((x) => {
-                        return x.name === 'CollectionsResolver';
-                    }),
+                    feedData: CollectionsResolver,
                 },
                 runGuardsAndResolvers: 'always',
             },
             {
                 path: 'collection/:collId',
-                component: PortfolioCollectionsPages.find((x) => {
-                    return x.name === 'CollectionPageComponent';
-                }),
+                component: CollectionPageComponent,
                 resolve: {
-                    collData: PortfolioResolvers.find((x) => {
-                        return x.name === 'CollectionPageResolver';
-                    }),
+                    collData: CollectionPageResolver,
                 },
                 runGuardsAndResolvers: 'always',
             },
             {
                 path: 'history',
-                component: PortfolioHistoryPages.find((x) => {
-                    return x.name === 'HistoryComponent';
-                }),
+                component: PortHistComponent,
                 canActivate: [AuthGuard],
                 resolve: {
-                    histData: PortfolioResolvers.find((x) => {
-                        return x.name === 'HistoryResolver';
-                    }),
+                    histData: HistoryResolver,
                 },
                 runGuardsAndResolvers: 'always',
             },
             {
                 path: 'conversations',
-                component: PortfolioConversationsPages.find((x) => {
-                    return x.name === 'ConversationsComponent';
-                }),
+                component: PortConversationsComponent,
                 canActivate: [AuthGuard],
             },
             {
                 path: 'settings',
-                component: PortfolioSettingsPages.find((x) => {
-                    return x.name === 'SettingsComponent';
-                }),
+                component: SettingsComponent,
                 canActivate: [AuthGuard],
             },
             {
                 path: 'home',
-                component: PortfolioHomePages.find((x) => {
-                    return x.name === 'PortHomeComponent';
-                }),
+                component: PortHomeComponent,
             },
             { path: '', redirectTo: 'home', pathMatch: 'full' },
         ],
