@@ -1,3 +1,1 @@
-import { PortHomeComponent } from './port-home.component';
-
-export const PortfolioHomePages = [PortHomeComponent];
+export { PortHomeComponent } from './port-home.component';
