@@ -1,3 +1,1 @@
-import { WorksComponent } from './works.component';
-
-export const PortfolioWorksPages = [WorksComponent];
+export { WorksComponent } from './works.component';
