@@ -1,4 +1,5 @@
-import { BrowseResolver } from './browse.resolver';
+export { ContentViewResolver } from './content-view.resolver';
+
 import { ContentViewResolver } from './content-view.resolver';
 
-export const Resolvers = [BrowseResolver, ContentViewResolver];
+export const Resolvers = [ContentViewResolver];
