@@ -22,16 +22,16 @@ async function bootstrap() {
     const app = await NestFactory.create(AppModule);
 
     // Initializing core functionality & security
+    app.use(helmet({ contentSecurityPolicy: false }));
     app.use(cookieParser(process.env.COOKIE_SECRET));
     app.use(json({ limit: '50mb' }));
     app.use(urlencoded({ limit: '50mb', extended: true }));
-    app.use(csurf({ cookie: true }));
+    app.use(csurf({ cookie: { sameSite: 'strict', secure: true }}));
     // this is to make sure the XSRF-TOKEN is being set correctly
     app.use(function (req, res, next) {
-        res.cookie('XSRF-TOKEN', req.csrfToken());
+        res.cookie('XSRF-TOKEN', req.csrfToken(), { sameSite: 'strict', secure: true });
         return next();
     });
-    app.use(helmet());
     app.use(require('prerender-node').set('prerenderToken', process.env.PRERENDER_TOKEN));
     app.setGlobalPrefix('api');
 
